+++ offer/src/main/java/com/cleevio/vexl/module/offer/service/OfferPublicRepository.java
@@ -17,13 +17,13 @@ interface OfferPublicRepository extends JpaRepository<OfferPublicPart, Long>, Jp
 
     @Modifying
     @Query("delete from OfferPublicPart o where o.adminId in :adminIds ")
-    long deleteByAdminIds(List<String> adminIds);
+    int deleteByAdminIds(List<String> adminIds);
 
     Optional<OfferPublicPart> findByAdminId(String adminId);
 
     @Modifying
     @Query("delete from OfferPublicPart p where p.refreshedAt < :expiration")
-    long deleteAllExpiredPublicParts(LocalDate expiration);
+    int deleteAllExpiredPublicParts(LocalDate expiration);
 
     @Query("select count(p) from OfferPublicPart p where p.offerType = :type and p.modifiedAt > :period ")
     int getModifiedOffersCount(LocalDate period, OfferType type);
